Validation and rest of the menu added

+++ App.UI/Menu/CompanyMenu.cs
@@ -12,15 +12,18 @@ namespace App.UI.Menu
     {
         private readonly ICompanyRepository _repository;
         private readonly ITakeInput _takeInput;
+        private readonly IPrintTables _printTables;
 
-        public CompanyMenu(ICompanyRepository repository, ITakeInput takeInput)
+        public CompanyMenu(ICompanyRepository repository, ITakeInput takeInput, IPrintTables printTables)
         {
             _repository = repository;
             _takeInput = takeInput;
+            _printTables = printTables;
         }
 
         public void Initialize()
         {
+
             MenuText();
             MenuSwitch();
         }
@@ -42,18 +45,16 @@ namespace App.UI.Menu
             switch (option)
             {
                 case (1):
-                    var companies = _repository.GetAll();
-                    var companiesTable = new ConsoleTable("Id", "Name", "Country", "Adress");
-                    foreach (var comp in companies)
-                    {
-                        companiesTable.AddRow(comp.Id, comp.Name, comp.Country, comp.Adress);
-                    }
-                    companiesTable.Write();
-                    Console.WriteLine();
+                    Console.Clear();
+
+                    _printTables.CompaniesTable();
+
                     Initialize();
                     break;
 
                 case (2):
+                    Console.Clear();
+
                     Console.Write("Podaj nazwę firmy: ");
                     var name = _takeInput.StringInput();
 
@@ -64,13 +65,23 @@ namespace App.UI.Menu
                     var adress = _takeInput.StringInput();
 
                     _repository.Create(name, country, adress);
+                    Console.Clear();
+
                     Initialize();
                     break;
 
                 case (3):
+                    Console.Clear();
+                    _printTables.CompaniesTable();
+
                     Console.Write("Podaj id firmy do usunięcia: ");
                     var idToDel = _takeInput.IntInput();
-                    _repository.Delete(idToDel);
+
+                    Console.WriteLine("Aby usunąć wprowadź 1, aby anulować wprowadź dowolną liczbę");
+                    var input = _takeInput.IntInput();
+                    if (input == 1)
+                        _repository.Delete(idToDel);
+                    Console.Clear();
                     Initialize();
                     break;
